+++ allo/ir/transform.py
@@ -300,162 +300,162 @@ def update_streaming_interface(module, target, depth=-1):
             func.attributes["function_type"] = TypeAttr.get(func_type)
 
 
-def create_buffer_load(tensors, name, ip, flatten=False, mapping=None):
+# def create_buffer_load(tensors, name, ip, flatten=False, mapping=None):
 
-    if len(tensors) != 2:
-        raise IndexError("One origin and one destination ONLY!")
+#     if len(tensors) != 2:
+#         raise IndexError("One origin and one destination ONLY!")
 
-    tensor_ori = tensors[0]
-    tensor_des = tensors[1]
+#     tensor_ori = tensors[0]
+#     tensor_des = tensors[1]
 
-    shape_des = MemRefType(tensor_des.type).shape
-    if mapping is not None:
-        loop_bounds, src_pattern, dst_pattern = mapping
-    else:
-        loop_bounds, src_pattern, dst_pattern = shape_des, None, None
+#     shape_des = MemRefType(tensor_des.type).shape
+#     if mapping is not None:
+#         loop_bounds, src_pattern, dst_pattern = mapping
+#     else:
+#         loop_bounds, src_pattern, dst_pattern = shape_des, None, None
 
-    for_loops = build_for_loops(loop_bounds, ip, name)
-    for_loops[-1].attributes["pipeline_ii"] = IntegerAttr.get(
-        IntegerType.get_unsigned(32), 1
-    )
-    for_loops[-1].attributes["rewind"] = UnitAttr.get()
-    induction_vars = [for_loop.induction_variable for for_loop in for_loops]
+#     for_loops = build_for_loops(loop_bounds, ip, name)
+#     for_loops[-1].attributes["pipeline_ii"] = IntegerAttr.get(
+#         IntegerType.get_unsigned(32), 1
+#     )
+#     for_loops[-1].attributes["rewind"] = UnitAttr.get()
+#     induction_vars = [for_loop.induction_variable for for_loop in for_loops]
 
-    with InsertionPoint(for_loops[-1].body.operations[0]):
-        if not flatten:
-            var_str = ", ".join([f"d{i}" for i in range(len(loop_bounds))])
-            if dst_pattern is None:
-                dst_pattern = var_str
-            affine_attr = AffineMapAttr.parse(
-                f"affine_map<({var_str})->({dst_pattern})>"
-            )
-            load = affine_d.AffineLoadOp(
-                MemRefType(tensor_ori.type).element_type,
-                tensor_ori,
-                induction_vars,
-                affine_attr,
-            )
-            affine_d.AffineStoreOp(
-                load.result,
-                tensor_des,
-                induction_vars,
-                affine_attr,
-            )
-        else:
-            # Build affine attribution
-            out_str = ""
-            reversed_shape = list(shape_des)[::-1]
-            for i in range(len(shape_des)):
-                s_str = " * ".join([str(s) for s in reversed_shape[:i]])
-                if s_str != "":
-                    out_str = s_str + f" * d{len(shape_des) - i - 1}" + out_str
-                else:
-                    out_str = f" d{len(shape_des) - i - 1}" + out_str
-                if i != len(shape_des) - 1:
-                    out_str = " + " + out_str
+#     with InsertionPoint(for_loops[-1].body.operations[0]):
+#         if not flatten:
+#             var_str = ", ".join([f"d{i}" for i in range(len(loop_bounds))])
+#             if dst_pattern is None:
+#                 dst_pattern = var_str
+#             affine_attr = AffineMapAttr.parse(
+#                 f"affine_map<({var_str})->({dst_pattern})>"
+#             )
+#             load = affine_d.AffineLoadOp(
+#                 MemRefType(tensor_ori.type).element_type,
+#                 tensor_ori,
+#                 induction_vars,
+#                 affine_attr,
+#             )
+#             affine_d.AffineStoreOp(
+#                 load.result,
+#                 tensor_des,
+#                 induction_vars,
+#                 affine_attr,
+#             )
+#         else:
+#             # Build affine attribution
+#             out_str = ""
+#             reversed_shape = list(shape_des)[::-1]
+#             for i in range(len(shape_des)):
+#                 s_str = " * ".join([str(s) for s in reversed_shape[:i]])
+#                 if s_str != "":
+#                     out_str = s_str + f" * d{len(shape_des) - i - 1}" + out_str
+#                 else:
+#                     out_str = f" d{len(shape_des) - i - 1}" + out_str
+#                 if i != len(shape_des) - 1:
+#                     out_str = " + " + out_str
 
-            in_str = ", ".join([f"d{i}" for i in range(len(loop_bounds))])
+#             in_str = ", ".join([f"d{i}" for i in range(len(loop_bounds))])
 
-            # Build LoadOp
-            if src_pattern is not None:
-                out_str = src_pattern
-            affine_attr = AffineMapAttr.parse(f"affine_map<({in_str})->({out_str})>")
-            load = affine_d.AffineLoadOp(
-                MemRefType(tensor_ori.type).element_type,
-                tensor_ori,
-                induction_vars,
-                affine_attr,
-            )
+#             # Build LoadOp
+#             if src_pattern is not None:
+#                 out_str = src_pattern
+#             affine_attr = AffineMapAttr.parse(f"affine_map<({in_str})->({out_str})>")
+#             load = affine_d.AffineLoadOp(
+#                 MemRefType(tensor_ori.type).element_type,
+#                 tensor_ori,
+#                 induction_vars,
+#                 affine_attr,
+#             )
 
-            # Build StoreOp
-            if dst_pattern is not None:
-                out_str = dst_pattern
-            else:
-                out_str = in_str
-            affine_attr = AffineMapAttr.parse(f"affine_map<({in_str})->({out_str})>")
-            affine_d.AffineStoreOp(
-                load.result,
-                tensor_des,
-                induction_vars,
-                affine_attr,
-            )
+#             # Build StoreOp
+#             if dst_pattern is not None:
+#                 out_str = dst_pattern
+#             else:
+#                 out_str = in_str
+#             affine_attr = AffineMapAttr.parse(f"affine_map<({in_str})->({out_str})>")
+#             affine_d.AffineStoreOp(
+#                 load.result,
+#                 tensor_des,
+#                 induction_vars,
+#                 affine_attr,
+#             )
 
 
-def create_buffer_store(tensors, name, ip, flatten=False, mapping=None):
+# def create_buffer_store(tensors, name, ip, flatten=False, mapping=None):
 
-    if len(tensors) != 2:
-        raise IndexError("One origin and one destination ONLY!")
+#     if len(tensors) != 2:
+#         raise IndexError("One origin and one destination ONLY!")
 
-    tensor_ori = tensors[0]
-    tensor_des = tensors[1]
+#     tensor_ori = tensors[0]
+#     tensor_des = tensors[1]
 
-    shape_ori = MemRefType(tensor_ori.type).shape
-    if mapping is not None:
-        loop_bounds, src_pattern, dst_pattern = mapping
-    else:
-        loop_bounds, src_pattern, dst_pattern = shape_ori, None, None
+#     shape_ori = MemRefType(tensor_ori.type).shape
+#     if mapping is not None:
+#         loop_bounds, src_pattern, dst_pattern = mapping
+#     else:
+#         loop_bounds, src_pattern, dst_pattern = shape_ori, None, None
 
-    for_loops = build_for_loops(loop_bounds, ip, name)
-    for_loops[-1].attributes["pipeline_ii"] = IntegerAttr.get(
-        IntegerType.get_unsigned(32), 1
-    )
-    for_loops[-1].attributes["rewind"] = UnitAttr.get()
-    induction_vars = [for_loop.induction_variable for for_loop in for_loops]
+#     for_loops = build_for_loops(loop_bounds, ip, name)
+#     for_loops[-1].attributes["pipeline_ii"] = IntegerAttr.get(
+#         IntegerType.get_unsigned(32), 1
+#     )
+#     for_loops[-1].attributes["rewind"] = UnitAttr.get()
+#     induction_vars = [for_loop.induction_variable for for_loop in for_loops]
 
-    with InsertionPoint(for_loops[-1].body.operations[0]):
-        if not flatten:
-            var_str = ", ".join([f"d{i}" for i in range(len(loop_bounds))])
-            if dst_pattern is None:
-                dst_pattern = var_str
-            affine_attr = AffineMapAttr.parse(
-                f"affine_map<({var_str})->({dst_pattern})>"
-            )
-            load = affine_d.AffineLoadOp(
-                MemRefType(tensor_ori.type).element_type,
-                tensor_ori,
-                induction_vars,
-                affine_attr,
-            )
-            affine_d.AffineStoreOp(
-                load.result,
-                tensor_des,
-                induction_vars,
-                affine_attr,
-            )
+#     with InsertionPoint(for_loops[-1].body.operations[0]):
+#         if not flatten:
+#             var_str = ", ".join([f"d{i}" for i in range(len(loop_bounds))])
+#             if dst_pattern is None:
+#                 dst_pattern = var_str
+#             affine_attr = AffineMapAttr.parse(
+#                 f"affine_map<({var_str})->({dst_pattern})>"
+#             )
+#             load = affine_d.AffineLoadOp(
+#                 MemRefType(tensor_ori.type).element_type,
+#                 tensor_ori,
+#                 induction_vars,
+#                 affine_attr,
+#             )
+#             affine_d.AffineStoreOp(
+#                 load.result,
+#                 tensor_des,
+#                 induction_vars,
+#                 affine_attr,
+#             )
 
-        else:
-            out_str = ""
-            reversed_shape = list(shape_ori)[::-1]
-            for i in range(len(shape_ori)):
-                s_str = " * ".join([str(s) for s in reversed_shape[:i]])
-                if s_str != "":
-                    out_str = s_str + f" * d{len(shape_ori) - i - 1}" + out_str
-                else:
-                    out_str = f" d{len(shape_ori) - i - 1}" + out_str
-                if i != len(shape_ori) - 1:
-                    out_str = " + " + out_str
+#         else:
+#             out_str = ""
+#             reversed_shape = list(shape_ori)[::-1]
+#             for i in range(len(shape_ori)):
+#                 s_str = " * ".join([str(s) for s in reversed_shape[:i]])
+#                 if s_str != "":
+#                     out_str = s_str + f" * d{len(shape_ori) - i - 1}" + out_str
+#                 else:
+#                     out_str = f" d{len(shape_ori) - i - 1}" + out_str
+#                 if i != len(shape_ori) - 1:
+#                     out_str = " + " + out_str
 
-            in_str = ", ".join([f"d{i}" for i in range(len(loop_bounds))])
-            if src_pattern is not None:
-                load_str = src_pattern
-            else:
-                load_str = in_str
-            affine_attr = AffineMapAttr.parse(f"affine_map<({in_str})->({load_str})>")
-            load = affine_d.AffineLoadOp(
-                MemRefType(tensor_ori.type).element_type,
-                tensor_ori,
-                induction_vars,
-                affine_attr,
-            )
-            if dst_pattern is not None:
-                out_str = dst_pattern
-            affine_attr = AffineMapAttr.parse(f"affine_map<({in_str})->({out_str})>")
-            affine_d.AffineStoreOp(
-                load.result,
-                tensor_des,
-                induction_vars,
-                affine_attr,
-            )
+#             in_str = ", ".join([f"d{i}" for i in range(len(loop_bounds))])
+#             if src_pattern is not None:
+#                 load_str = src_pattern
+#             else:
+#                 load_str = in_str
+#             affine_attr = AffineMapAttr.parse(f"affine_map<({in_str})->({load_str})>")
+#             load = affine_d.AffineLoadOp(
+#                 MemRefType(tensor_ori.type).element_type,
+#                 tensor_ori,
+#                 induction_vars,
+#                 affine_attr,
+#             )
+#             if dst_pattern is not None:
+#                 out_str = dst_pattern
+#             affine_attr = AffineMapAttr.parse(f"affine_map<({in_str})->({out_str})>")
+#             affine_d.AffineStoreOp(
+#                 load.result,
+#                 tensor_des,
+#                 induction_vars,
+#                 affine_attr,
+#             )
 
 
 def create_data_movement(
@@ -590,163 +590,163 @@ def wrap_data_movement(arg, ip, func_name, from_memory, flatten, mapping):
     func_d.ReturnOp([], ip=InsertionPoint(func_op.entry_block))
 
 
-def create_buffer(tensor, name, ip, alloc_ip=None, flatten=False, mapping=None):
-    new_ip = ip if alloc_ip is None else alloc_ip
-    if not isinstance(new_ip, InsertionPoint):
-        new_ip = InsertionPoint(new_ip)
-    with new_ip:
-        shape = MemRefType(tensor.type).shape
-        if not flatten or alloc_ip is None:
-            alloc_op = memref_d.AllocOp(
-                # remove layout & memory_space info
-                MemRefType.get(
-                    MemRefType(tensor.type).shape, MemRefType(tensor.type).element_type
-                ),
-                [],
-                [],
-            )
-        else:  # store back to results
-            alloc_op = memref_d.AllocOp(
-                MemRefType.get((np.prod(shape),), MemRefType(tensor.type).element_type),
-                [],
-                [],
-            )
-        alloc_op.attributes["name"] = StringAttr.get(name)
-    if alloc_ip is None:  # load
-        tensor.replace_all_uses_with(alloc_op.result)
-    if mapping is not None:
-        loop_bounds, src_pattern, dst_pattern = mapping
-    else:
-        loop_bounds, src_pattern, dst_pattern = shape, None, None
+# def create_buffer(tensor, name, ip, alloc_ip=None, flatten=False, mapping=None):
+#     new_ip = ip if alloc_ip is None else alloc_ip
+#     if not isinstance(new_ip, InsertionPoint):
+#         new_ip = InsertionPoint(new_ip)
+#     with new_ip:
+#         shape = MemRefType(tensor.type).shape
+#         if not flatten or alloc_ip is None:
+#             alloc_op = memref_d.AllocOp(
+#                 # remove layout & memory_space info
+#                 MemRefType.get(
+#                     MemRefType(tensor.type).shape, MemRefType(tensor.type).element_type
+#                 ),
+#                 [],
+#                 [],
+#             )
+#         else:  # store back to results
+#             alloc_op = memref_d.AllocOp(
+#                 MemRefType.get((np.prod(shape),), MemRefType(tensor.type).element_type),
+#                 [],
+#                 [],
+#             )
+#         alloc_op.attributes["name"] = StringAttr.get(name)
+#     if alloc_ip is None:  # load
+#         tensor.replace_all_uses_with(alloc_op.result)
+#     if mapping is not None:
+#         loop_bounds, src_pattern, dst_pattern = mapping
+#     else:
+#         loop_bounds, src_pattern, dst_pattern = shape, None, None
 
-    for_loops = build_for_loops(loop_bounds, ip, name)
-    for_loops[-1].attributes["pipeline_ii"] = IntegerAttr.get(
-        IntegerType.get_unsigned(32), 1
-    )
-    for_loops[-1].attributes["rewind"] = UnitAttr.get()
-    induction_vars = [for_loop.induction_variable for for_loop in for_loops]
-    with InsertionPoint(for_loops[-1].body.operations[0]):
-        if not flatten:
-            var_str = ", ".join([f"d{i}" for i in range(len(loop_bounds))])
-            if dst_pattern is None:
-                dst_pattern = var_str
-            affine_attr = AffineMapAttr.parse(
-                f"affine_map<({var_str})->({dst_pattern})>"
-            )
-            load = affine_d.AffineLoadOp(
-                MemRefType(tensor.type).element_type,
-                tensor,
-                induction_vars,
-                affine_attr,
-            )
-            affine_d.AffineStoreOp(
-                load.result,
-                alloc_op.result,
-                induction_vars,
-                affine_attr,
-            )
-        else:
-            out_str = ""
-            reversed_shape = list(shape)[::-1]
-            for i in range(len(shape)):
-                s_str = " * ".join([str(s) for s in reversed_shape[:i]])
-                if s_str != "":
-                    out_str = s_str + f" * d{len(shape) - i - 1}" + out_str
-                else:
-                    out_str = f" d{len(shape) - i - 1}" + out_str
-                if i != len(shape) - 1:
-                    out_str = " + " + out_str
-            if alloc_ip is None:  # load from inputs
-                in_str = ", ".join([f"d{i}" for i in range(len(loop_bounds))])
-                if src_pattern is not None:
-                    out_str = src_pattern
-                affine_attr = AffineMapAttr.parse(
-                    f"affine_map<({in_str})->({out_str})>"
-                )
-                load = affine_d.AffineLoadOp(
-                    MemRefType(tensor.type).element_type,
-                    tensor,
-                    induction_vars,
-                    affine_attr,
-                )
-                if dst_pattern is not None:
-                    out_str = dst_pattern
-                else:
-                    out_str = in_str
-                affine_attr = AffineMapAttr.parse(
-                    f"affine_map<({in_str})->({out_str})>"
-                )
-                affine_d.AffineStoreOp(
-                    load.result,
-                    alloc_op.result,
-                    induction_vars,
-                    affine_attr,
-                )
-            else:  # store back results to outputs
-                in_str = ", ".join([f"d{i}" for i in range(len(loop_bounds))])
-                if src_pattern is not None:
-                    load_str = src_pattern
-                else:
-                    load_str = in_str
-                affine_attr = AffineMapAttr.parse(
-                    f"affine_map<({in_str})->({load_str})>"
-                )
-                load = affine_d.AffineLoadOp(
-                    MemRefType(tensor.type).element_type,
-                    tensor,
-                    induction_vars,
-                    affine_attr,
-                )
-                if dst_pattern is not None:
-                    out_str = dst_pattern
-                affine_attr = AffineMapAttr.parse(
-                    f"affine_map<({in_str})->({out_str})>"
-                )
-                affine_d.AffineStoreOp(
-                    load.result,
-                    alloc_op.result,
-                    induction_vars,
-                    affine_attr,
-                )
-    return alloc_op
+#     for_loops = build_for_loops(loop_bounds, ip, name)
+#     for_loops[-1].attributes["pipeline_ii"] = IntegerAttr.get(
+#         IntegerType.get_unsigned(32), 1
+#     )
+#     for_loops[-1].attributes["rewind"] = UnitAttr.get()
+#     induction_vars = [for_loop.induction_variable for for_loop in for_loops]
+#     with InsertionPoint(for_loops[-1].body.operations[0]):
+#         if not flatten:
+#             var_str = ", ".join([f"d{i}" for i in range(len(loop_bounds))])
+#             if dst_pattern is None:
+#                 dst_pattern = var_str
+#             affine_attr = AffineMapAttr.parse(
+#                 f"affine_map<({var_str})->({dst_pattern})>"
+#             )
+#             load = affine_d.AffineLoadOp(
+#                 MemRefType(tensor.type).element_type,
+#                 tensor,
+#                 induction_vars,
+#                 affine_attr,
+#             )
+#             affine_d.AffineStoreOp(
+#                 load.result,
+#                 alloc_op.result,
+#                 induction_vars,
+#                 affine_attr,
+#             )
+#         else:
+#             out_str = ""
+#             reversed_shape = list(shape)[::-1]
+#             for i in range(len(shape)):
+#                 s_str = " * ".join([str(s) for s in reversed_shape[:i]])
+#                 if s_str != "":
+#                     out_str = s_str + f" * d{len(shape) - i - 1}" + out_str
+#                 else:
+#                     out_str = f" d{len(shape) - i - 1}" + out_str
+#                 if i != len(shape) - 1:
+#                     out_str = " + " + out_str
+#             if alloc_ip is None:  # load from inputs
+#                 in_str = ", ".join([f"d{i}" for i in range(len(loop_bounds))])
+#                 if src_pattern is not None:
+#                     out_str = src_pattern
+#                 affine_attr = AffineMapAttr.parse(
+#                     f"affine_map<({in_str})->({out_str})>"
+#                 )
+#                 load = affine_d.AffineLoadOp(
+#                     MemRefType(tensor.type).element_type,
+#                     tensor,
+#                     induction_vars,
+#                     affine_attr,
+#                 )
+#                 if dst_pattern is not None:
+#                     out_str = dst_pattern
+#                 else:
+#                     out_str = in_str
+#                 affine_attr = AffineMapAttr.parse(
+#                     f"affine_map<({in_str})->({out_str})>"
+#                 )
+#                 affine_d.AffineStoreOp(
+#                     load.result,
+#                     alloc_op.result,
+#                     induction_vars,
+#                     affine_attr,
+#                 )
+#             else:  # store back results to outputs
+#                 in_str = ", ".join([f"d{i}" for i in range(len(loop_bounds))])
+#                 if src_pattern is not None:
+#                     load_str = src_pattern
+#                 else:
+#                     load_str = in_str
+#                 affine_attr = AffineMapAttr.parse(
+#                     f"affine_map<({in_str})->({load_str})>"
+#                 )
+#                 load = affine_d.AffineLoadOp(
+#                     MemRefType(tensor.type).element_type,
+#                     tensor,
+#                     induction_vars,
+#                     affine_attr,
+#                 )
+#                 if dst_pattern is not None:
+#                     out_str = dst_pattern
+#                 affine_attr = AffineMapAttr.parse(
+#                     f"affine_map<({in_str})->({out_str})>"
+#                 )
+#                 affine_d.AffineStoreOp(
+#                     load.result,
+#                     alloc_op.result,
+#                     induction_vars,
+#                     affine_attr,
+#                 )
+#     return alloc_op
 
 
-def store_tensor(tensor, target, name, ip, flatten=True):
-    shape = MemRefType(tensor.type).shape
-    loop_bounds = shape
-    for_loops = build_for_loops(loop_bounds, ip, name)
-    for_loops[-1].attributes["pipeline_ii"] = IntegerAttr.get(
-        IntegerType.get_unsigned(32), 1
-    )
-    for_loops[-1].attributes["rewind"] = UnitAttr.get()
-    induction_vars = [for_loop.induction_variable for for_loop in for_loops]
-    with InsertionPoint(for_loops[-1].body.operations[0]):
-        in_str = ", ".join([f"d{i}" for i in range(len(loop_bounds))])
-        load_str = in_str
-        affine_attr = AffineMapAttr.parse(f"affine_map<({in_str})->({load_str})>")
-        load = affine_d.AffineLoadOp(
-            MemRefType(tensor.type).element_type, tensor, induction_vars, affine_attr
-        )
-        if not flatten:
-            out_str = in_str
-        else:
-            out_str = ""
-            reversed_shape = list(shape)[::-1]
-            for i in range(len(shape)):
-                s_str = " * ".join([str(s) for s in reversed_shape[:i]])
-                if s_str != "":
-                    out_str = s_str + f" * d{len(shape) - i - 1}" + out_str
-                else:
-                    out_str = f" d{len(shape) - i - 1}" + out_str
-                if i != len(shape) - 1:
-                    out_str = " + " + out_str
-        affine_attr = AffineMapAttr.parse(f"affine_map<({in_str})->({out_str})>")
-        affine_d.AffineStoreOp(
-            load.result,
-            target,
-            induction_vars,
-            affine_attr,
-        )
+# def store_tensor(tensor, target, name, ip, flatten=True):
+#     shape = MemRefType(tensor.type).shape
+#     loop_bounds = shape
+#     for_loops = build_for_loops(loop_bounds, ip, name)
+#     for_loops[-1].attributes["pipeline_ii"] = IntegerAttr.get(
+#         IntegerType.get_unsigned(32), 1
+#     )
+#     for_loops[-1].attributes["rewind"] = UnitAttr.get()
+#     induction_vars = [for_loop.induction_variable for for_loop in for_loops]
+#     with InsertionPoint(for_loops[-1].body.operations[0]):
+#         in_str = ", ".join([f"d{i}" for i in range(len(loop_bounds))])
+#         load_str = in_str
+#         affine_attr = AffineMapAttr.parse(f"affine_map<({in_str})->({load_str})>")
+#         load = affine_d.AffineLoadOp(
+#             MemRefType(tensor.type).element_type, tensor, induction_vars, affine_attr
+#         )
+#         if not flatten:
+#             out_str = in_str
+#         else:
+#             out_str = ""
+#             reversed_shape = list(shape)[::-1]
+#             for i in range(len(shape)):
+#                 s_str = " * ".join([str(s) for s in reversed_shape[:i]])
+#                 if s_str != "":
+#                     out_str = s_str + f" * d{len(shape) - i - 1}" + out_str
+#                 else:
+#                     out_str = f" d{len(shape) - i - 1}" + out_str
+#                 if i != len(shape) - 1:
+#                     out_str = " + " + out_str
+#         affine_attr = AffineMapAttr.parse(f"affine_map<({in_str})->({out_str})>")
+#         affine_d.AffineStoreOp(
+#             load.result,
+#             target,
+#             induction_vars,
+#             affine_attr,
+#         )
 
 
 def find_func_in_module(module, func_name):
