+++ tests/test_vhls.py
@@ -18,7 +18,9 @@ def test_io_buffer_gemm(flatten):
 
     s = allo.customize(gemm)
     print(s.module)
-    allo.passes.generate_input_output_buffers(s.top_func, flatten=flatten)
+    allo.passes.generate_input_output_buffers(
+        s.module, s.top_func_name, flatten=flatten
+    )
     print(s.module)
     mod = s.build()
     if not flatten:
@@ -100,7 +102,9 @@ def test_vitis_io_stream():
     if hls.is_available("vitis_hls"):
         hls_mod = s.build(target="vitis_hls", mode="sw_emu", project="test_io.prj")
         print(s.module)
-        hls_mod()
+        np_A = np.random.randint(0, 10, size=(32, 32)).astype(np.int32)
+        hls_mod(np_A)
+        # hls_mod()
 
 
 def test_csim_write_back():
@@ -127,7 +131,7 @@ def test_pointer_generation():
 
     s = allo.customize(top)
     mod = s.build(target="vitis_hls", mode="csim", project="test_pointer.prj")
-    assert "bool v0," in mod.hls_code and ",," not in mod.hls_code
+    assert "bool v" in mod.hls_code and ",," not in mod.hls_code
     if hls.is_available("vitis_hls"):
         inst = np.array([1], dtype=np.bool_)
         C = np.array([1, 2, 3], dtype=np.int32)
@@ -144,7 +148,7 @@ def test_scalar_not_array():
 
     s = allo.customize(top)
     mod = s.build(target="vitis_hls", mode="csim", project="test_scalar.prj")
-    assert "bool v0," in mod.hls_code and ",," not in mod.hls_code
+    assert "bool v" in mod.hls_code and ",," not in mod.hls_code
     if hls.is_available("vitis_hls"):
         C = np.array([1, 2, 3], dtype=np.int32)
         mod(1, C)
@@ -182,6 +186,26 @@ def test_size1_array():
         np_B = np.array([0], dtype=np.int32)
         mod(np_A, np_B)
         np.testing.assert_allclose(np_A, [2], rtol=1e-5)
+        print("Passed!")
+
+
+def test_wrap_nonvoid():
+    M, N = 4, 4
+
+    def matrix_add(A: float32[M, N]) -> float32[M, N]:
+        B: float32[M, N]
+        for i, j in allo.grid(M, N, name="PE"):
+            B[i, j] = A[i, j] + 1
+        return B
+
+    s = allo.customize(matrix_add)
+
+    if hls.is_available("vitis_hls"):
+        mod = s.build(target="vitis_hls", mode="csim")
+        module = str(mod.module)
+        assert (
+            f"func.func @matrix_add(%arg0: memref<16xf32>) -> memref<16xf32>" in module
+        )
         print("Passed!")
